requirements.txt
streamlit==1.52.1
Pillow==12.0.0
oracledb==3.4.1
